requirements.txt
graphviz==0.8
numpy==1.14.0
Pillow==4.3.0
tqdm==4.19.5
scipy==1.0.0
torch==0.3.1.post2
torchvision==0.2.0
pretrainedmodels==0.6.0
matplotlib==2.1.0
h5py==2.7.1
importlib_resources==0.4.0
semantic_version==2.6.0
sacred==0.7.3
pymongo==3.6.0
GitPython==2.1.8
tabulate==0.8.2
plotly==2.3.0
seaborn==0.8.1
git+https://github.com/anibali/pyshowoff.git@3396cb916c390886873fb0a6b8515f91c0c1574c#egg=pyshowoff
git+https://github.com/anibali/tele.git@e2b0af5f28100abd581dbf3871bee90b178a5c98#egg=tele
